chore: move Menu, MenuGroup to model package
